requirements.txt
discord.py
python-dotenv
asyncpg
fastapi
uvicorn
pillow
static-ffmpeg
pynacl
yt_dlp
async_timeout
apscheduler
emoji
wordfreq
deep_translator
gtts
dotenv
torch
psutil
transformers
accelerate
hf_xet
textblob
langdetect
optimum 
llama-cpp-python
rapidfuzz
matplotlib
topggpy
pycountry
rapidfuzz
py-evalexpr
groq